fixed problem with the order beans were getting loaded
